Add Blake3SubtreeHasher for feeding a piece in parts; allow pieces over 2 GB
Blake3SubtreeContext.CreateSubtreeHasher(pieceIndex) returns an
incremental hasher: Update as the piece's bytes arrive, Finish to get
the same Blake3Subtree that HashSubtree produces. A piece no longer has
to be in memory at once, and the piece size is now a long, so pieces
larger than a span can hold are allowed (they must go through the
incremental hasher; HashSubtree says so when asked for one).

Internally HasherState takes a start chunk. Chunk counters stay absolute
while the merge bookkeeping counts chunks from the piece start, since
the pairing rule depends on the relative count and the absolute one
would merge past the top of the stack at the end of an aligned piece.

Breaking for 1.4.0 callers only in signatures: pieceSize parameters,
PieceSize, GetPieceLength and Blake3Subtree.Length are long now.
Source that passed ints still compiles.

Tests: 105 new cases feed pieces in random fragment sizes across seven
piece sizes and fourteen lengths, mix incremental and one-shot pieces,
cover keyed and derive-key modes, XOF, concurrency and every error path.
Serial Update() A/B is unchanged within noise.

Resolves the follow-up on GitHub issue #2.

+++ src/Blake3.Managed.Tests/Blake3SubtreeContextTests.cs
@@ -127,7 +127,7 @@ public class Blake3SubtreeContextTests
         Parallel.ForEach(Enumerable.Range(0, ctx.PieceCount).Reverse(), i =>
         {
             pieces[i] = ctx.HashSubtree(
-                data.AsSpan((int)ctx.GetPieceOffset(i), ctx.GetPieceLength(i)), i);
+                data.AsSpan((int)ctx.GetPieceOffset(i), (int)ctx.GetPieceLength(i)), i);
         });
 
         Assert.Equal(Hasher.Hash(data).ToString(), ctx.Finalize(pieces).ToString());
